test(diagnostics): respect layer and platform boundaries
